requirements.txt
# MongoDB + Ditto Conversion - Python Dependencies
# Phase 2: MongoDB Setup & Configuration

# MongoDB
pymongo>=4.6.0              # MongoDB driver
motor>=3.3.0                # Async MongoDB driver (for performance)

# Data Generation
faker>=20.0.0               # Fake data generation (customers, names, emails)

# Configuration
python-dotenv>=1.0.0        # Environment variables from .env
pyyaml>=6.0                 # YAML configuration files

# AI Embeddings
openai>=1.0.0               # OpenAI embeddings API
azure-identity>=1.15.0      # Azure auth for OpenAI
azure-ai-inference>=1.0.0b9 # Azure AI services (beta)

# Testing & Validation
pytest>=7.4.0               # Testing framework
pytest-asyncio>=0.21.0      # Async testing support

# Utilities
pytz>=2023.3                # Timezone handling
